+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * a new quarto vignette
 
+* AER, mlogit and plm are no longer suggested packages
+
 # dfidx 0.0-5
 
 * test of the presence of suggested packages are added in the man